requirements.txt
streamlit==1.43.2
langchain==0.3.23
langchain-google-genai==2.1.2
langchain-core== 0.3.51
langchain-community==0.3.21
huggingface-hub==0.30.1
requests== 2.32.3
python-dotenv==1.1.0
PyPDF2==3.0.1
transformers==4.50.3
langchain-openai==0.3.12
chromadb==1.0.4
google-generativeai=0.8.4

